修复处理分享率问题导致无法推送种子的 BUG

+++ lib/app/home/pages/agg_search/download_form.dart
@@ -307,10 +307,8 @@ class DownloadForm extends StatelessWidget {
                         isLoading.value = true;
                         double? ratioLimit =
                             double.tryParse(ratioLimitController.text);
-                        double? upLimit =
-                            double.tryParse(upLimitController.text);
-                        double? dlLimit =
-                            double.tryParse(dlLimitController.text);
+                        int? upLimit = int.tryParse(upLimitController.text);
+                        int? dlLimit = int.tryParse(dlLimitController.text);
                         await submitForm({
                           'mySite': mysite,
                           'magnet': urlController.text,
